requirements.txt
# requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.9
python-dotenv==1.0.1
requests==2.32.3